Split library database module
